refactor(ai): extract marker color normalization
